# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: LIMAPAY API - Gestion commerciale pour commerces physiques du Sénégal
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /sales:
    get:
      operationId: listSales
      tags: [sales]
      summary: Lister les ventes
      parameters:
        - name: dateFrom
          in: query
          schema:
            type: string
            format: date
        - name: dateTo
          in: query
          schema:
            type: string
            format: date
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Liste des ventes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Sale"
    post:
      operationId: createSale
      tags: [sales]
      summary: Créer une vente
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaleInput"
      responses:
        "201":
          description: Vente créée
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sale"
  /sales/{id}:
    get:
      operationId: getSale
      tags: [sales]
      summary: Obtenir une vente
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Vente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sale"
        "404":
          description: Vente introuvable

  # CUSTOMERS
...
components:
  schemas:
...
    SaleItem:
      type: object
      required: [productId, productName, quantity, unitPrice, totalPrice]
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: number
        totalPrice:
          type: number
    Sale:
      type: object
      required: [id, totalAmount, paymentMethod, status, items]
      properties:
        id:
          type: integer
        totalAmount:
          type: number
        amountPaid:
          type: number
        changeAmount:
          type: number
        paymentMethod:
          type: string
          enum: [cash, wave, orange_money, free_money, diamondpay, credit]
        status:
          type: string
          enum: [completed, pending, cancelled]
        customerId:
          type: ["integer", "null"]
        customerName:
          type: ["string", "null"]
        diamondPayTransactionId:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/SaleItem"
        createdAt:
          type: string
          format: date-time
    SaleItemInput:
      type: object
      required: [productId, quantity]
      properties:
        productId:
          type: integer
        quantity:
          type: integer
    SaleInput:
      type: object
      required: [items, paymentMethod, totalAmount]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/SaleItemInput"
        paymentMethod:
          type: string
          enum: [cash, wave, orange_money, free_money, diamondpay, credit]
        totalAmount:
          type: number
        amountPaid:
          type: number
        customerId:
          type: integer
        diamondPayTransactionId:
          type: string

    # CUSTOMERS
...